Refactor `multi_runner`
