+++ mle_toolbox/experiment/__init__.py
@@ -1,6 +1,12 @@
 from .experiment import Experiment
+from .spawn_single_job import (spawn_single_experiment,
+                               spawn_post_processing_job)
+from .spawn_multi_configs import spawn_multiple_configs
 
 
 __all__ = [
-           'Experiment'
+           'Experiment',
+           "spawn_single_experiment",
+           "spawn_post_processing_job",
+           "spawn_multiple_configs"
            ]
