+++ mle_toolbox/launch/single_experiment.py
@@ -1,5 +1,6 @@
 from ..experiment import Experiment
 
+
 def run_single_experiment(meta_job_args: dict,
                           single_job_args: dict):
     """ Run a single experiment locally/remote. """
