--- mle_toolbox/multi_runner/__init__.py
@@ -1,9 +1,0 @@
-from .spawn_single_run import (spawn_single_experiment,
-                               spawn_post_processing_job)
-from .spawn_multiple_runs import spawn_multiple_runs
-
-__all__ = [
-           "spawn_single_experiment",
-           "spawn_post_processing_job",
-           "spawn_multi_runs"
-          ]
